1. Changes to view dragging. Only done by title bar now (sort of stippled to show this - needs to be a bit clearer). Means map can now be dragged by title bar without interfering with slippy map controls.
2. Map updates to reflect interactions / style of other two main views (timeline/taxonomy).
Clusters are log scaled to the amount of stuff in a cluster, with partial shading in each cluster according to selected amount. Again this is log scaled, so a small selection is reflected
Right-click selects clusters and markers, same interaction as in timeline/taxonomy view
tooltip style now the same.